requirements.txt
# Core ML requirements 
numpy>=1.26.0
pandas>=2.1.0
scikit-learn>=1.3.2
xgboost>=2.0.0

# MLOps
mlflow==2.8.0
dvc==3.30.0
dvc-s3==3.0.1

# API & Serving
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2

# Monitoring
prometheus-client==0.18.0
psutil==5.9.6
evidently==0.4.10

# Utilities
pyyaml==6.0.1
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Visualization (optional)
matplotlib==3.7.3
seaborn==0.12.2
psycopg2-binary>=2.9.7
supabase
